move retweeter to apps folder
