+++ tests/apps/github_stats/cassettes/test_generate_summary.yaml
