+++ tests/apps/github_stats/cassettes/test_generates_empty_summary_if_no_events_found.yaml
