+++ tests/apps/github_stats/cassettes/test_post_github_summary_task__integration.yaml
